Switch rotation to quaternions; fix index buffer size and PlayerController typo
- Change GameObject::m_rotation from Euler-angle glm::vec3 to
  glm::quat; GetLocalTransform() composes it via glm::mat4_cast
- Rewrite PlayerControllerComponent's mouse-look for quaternions:
  yaw around world-up, pitch around local right, combined and
  renormalized; front/right now come from rotating unit axes by
  the current rotation
- Rework CameraComponent::GetViewMatrix() to hand-build a
  rotation+position matrix (excluding the camera's own scale)
  instead of inverting GetWorldTransform() directly
- Fix GraphicsAPI::CreateIndexBuffer uploading indices.size() *
  sizeof(float) instead of sizeof(uint32_t) (open since the first
  commit)
- Rename PlayerControllerConmponent -> PlayerControllerComponent

+++ engine/source/graphics/GraphicsAPI.cpp
@@ -122,7 +122,7 @@ namespace eng
         GLuint ebo = 0;
         glGenBuffers(1, &ebo);
         glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, ebo);
-        glBufferData(GL_ELEMENT_ARRAY_BUFFER, indices.size() * sizeof(float), indices.data(), GL_STATIC_DRAW);
+        glBufferData(GL_ELEMENT_ARRAY_BUFFER, indices.size() * sizeof(uint32_t), indices.data(), GL_STATIC_DRAW);
         glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
         return ebo;
     }
